handle req

+++ lib/src/dio_internet_interceptor.dart
@@ -19,9 +19,11 @@ class DioInternetInterceptor extends Interceptor {
   const DioInternetInterceptor({
     required this.onDioRequest,
     required this.onDioError,
+    this.hasConnection,
   });
 
   final RequestOptions Function(RequestOptions options) onDioRequest;
+  final void Function({bool isConnected})? hasConnection;
   final DioExeptionHander Function(
     DioException err,
     ErrorInterceptorHandler handler,
@@ -36,17 +38,22 @@ class DioInternetInterceptor extends Interceptor {
 
     final result = await InternetConnectionChecker().hasConnection;
     if (result == false) {
+      if (options.method.toUpperCase() == 'GET') {
+        hasConnection?.call(isConnected: result);
+        return;
+      }
       final curl = _cURLRepresentation(reqOptions);
       final curlService = CurlService();
       await curlService.addCurl(curl);
       final value = await curlService.getCurls();
-      /* value?.forEach((element) {
+     /* value?.forEach((element) {
         curlService.deleteCurl(value.indexOf(element));
       });*/
       print(value);
-
+      hasConnection?.call(isConnected: result);
       return;
     } else {
+      print(options.data);
       await _makeRequest();
       handler.next(reqOptions);
     }
@@ -72,12 +79,12 @@ class DioInternetInterceptor extends Interceptor {
         components.add('-H "$k: $v"');
       }
     });
-    print(components);
 
     if (options.data != null) {
       final data = json.encode(options.data).replaceAll('"', '\\"');
       components.add('-d "$data"');
     }
+
 
     components.add('"${options.uri.toString()}"');
 
@@ -86,7 +93,7 @@ class DioInternetInterceptor extends Interceptor {
 
   RequestOptions _curlRepresentationToOptions(String curl) {
     // Split the string by spaces to get individual parts
-    print(curl);
+
     final parts = curl.split(' ');
 
     // Extract method from the parts
@@ -109,13 +116,23 @@ class DioInternetInterceptor extends Interceptor {
 
     // Extract headers from the parts
     final headers = extractHeadersFromCurl(curl);
-    print(headers);
 
     // Extract body from the parts
     dynamic body;
     for (var i = 0; i < parts.length; i++) {
       if (parts[i] == '-d' && i + 1 < parts.length) {
-        body = json.decode(parts[i + 1].replaceAll('"', ''));
+        // Extract the data section (inside double quotes)
+        String dataSection = parts[i + 1];
+
+        // Remove surrounding double quotes and escape characters from the data section
+        dataSection = dataSection.replaceAll('"', '');
+
+        // Decode the JSON string into a Dart object
+        try {
+          body = json.decode(dataSection);
+        } catch (e) {
+          throw FormatException('Error parsing request body: $e');
+        }
         break;
       }
     }
@@ -137,6 +154,46 @@ class DioInternetInterceptor extends Interceptor {
       ..queryParameters = uri.queryParameters;
 
     return options;
+  }
+
+  Map<String, dynamic> extractBodysFromCurl(String curl) {
+    final bodyMap = <String, dynamic>{};
+
+    // Split the curl command by spaces to analyze each part
+    final parts = curl.split(' ');
+
+    for (var i = 0; i < parts.length; i++) {
+      // Look for the '-d' flag
+      if (parts[i] == '-d' && i + 1 < parts.length) {
+        // Extract the data section
+        final dataSection = parts[i + 1];
+
+        // Check if it starts with '{' and ends with '}', indicating a JSON object
+        if (dataSection.startsWith('{') && dataSection.endsWith('}')) {
+          try {
+            // Parse the JSON string into a map
+            final parsedData = json.decode(dataSection);
+
+            // Add each key-value pair from the parsed data into the bodyMap
+            parsedData.forEach((String key, value) {
+              String cleanedKey = key.replaceAll(r'\', '');
+              bodyMap["$cleanedKey "] = ("$value " as String).replaceAll(r'\', '');
+            });
+
+            // Break the loop since we found and parsed the data section
+            break;
+          } catch (e) {
+            // Handle any parsing errors here if needed
+            print('Error parsing data section: $e');
+          }
+        } else {
+          // Handle cases where the data section format is unexpected
+          print('Unexpected data section format: $dataSection');
+        }
+      }
+    }
+
+    return bodyMap;
   }
 
   Map<String, dynamic> extractHeadersFromCurl(String curl) {
@@ -216,7 +273,7 @@ class DioInternetInterceptor extends Interceptor {
         curlList.length,
         (index) => dio.request(
           curlList[0].keys.first,
-          data: curlList[0].values.first,
+          options: curlList[0].values.first as Options,
         ),
       ),
     );
